Fixed a bug in vcov.gnm (which gave an error when data are of class table).
Changed all the model-fitting tests to print out only the deviance (to 10 
digits) and degrees of freedom, rather than the full list of coefficients.
This should be more stable -- we are not really interested in the 
parameterization, but it is useful to know whether code changes affect
the model deviance and/or df. 


git-svn-id: file:///home/hturner/gnm@358 4c16e084-14e8-0310-9088-c52a9735df6d