Substituted \method{}{gnm} as necessary in Rd files.
Changed computation of thetaID in gnmTools - fixes problem of split blocks of terms but doesn't work for factors with common effects - need to fix.

Quick fix of names for matched factors - not thoroughly tested.

Evaluation of "nonlin" terms now in environment of data passed to gnm - sufficient for current use but probably not bullet proof.

Suppressed warnings in evaluation of local design matrix when theta allowed to be NA.

git-svn-id: file:///home/hturner/gnm@402 4c16e084-14e8-0310-9088-c52a9735df6d